fix money store type

+++ migrations/00003_create_balances_table.sql
@@ -2,8 +2,8 @@
 
 CREATE TABLE IF NOT EXISTS balances (
     user_id bigint not null PRIMARY KEY,
-    current double precision ,
-    withdrawn double precision ,
+    current decimal(19,2) ,
+    withdrawn decimal(19,2) ,
     changed_at timestamptz not null DEFAULT NOW()
 );
