+++ migrations/00004_create_withdrawn_table.sql
@@ -3,7 +3,7 @@
 CREATE TABLE IF NOT EXISTS withdrawals (
     user_id bigint,
     order_id bigint not null PRIMARY KEY,
-    sum double precision,
+    sum decimal(19,2),
     processed_at timestamptz not null DEFAULT NOW()
 );
